requirements.txt
emoji>=2.0.0
pyahocorasick>=2.0.0
tiktoken>=0.5.0
